[Fix] 필요없는 의존성 제거 예외처리 추가

+++ backend/src/main/java/com/ssafy/api/service/LiveServiceImpl.java
@@ -2,14 +2,12 @@ package com.ssafy.api.service;
 
 import com.ssafy.api.request.*;
 import com.ssafy.api.response.*;
-import com.ssafy.common.error.ErrorCode;
 import com.ssafy.common.exception.CustomException;
 import com.ssafy.common.util.DistanceModule;
 import com.ssafy.common.util.LocationDistance;
 import com.ssafy.db.entity.*;
 import com.ssafy.db.repository.*;
 import lombok.RequiredArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.*;
@@ -27,8 +25,7 @@ public class LiveServiceImpl implements LiveService {
     private final UserLiveRepository userLiveRepository;
     private final ProductRepository productRepository;
     private final FavoriteRepository favoriteRepository;
-    private final int MAX_PAGE = 10;
-
+    private static final int MAX_PAGE = 10;
 
     @Override
     public Live CreateLive(LiveRegisterPostReq liveRegisterInfo, User user) {
@@ -54,7 +51,7 @@ public class LiveServiceImpl implements LiveService {
 
         liveRepository.save(live);
 
-        //유저라이브 헬퍼 테이블에 본인도 넣어주기
+        // 유저라이브 헬퍼 테이블에 본인도 넣어주기
         UserLive userLive = UserLive.builder()
                 .live(live)
                 .user(user)
@@ -69,7 +66,7 @@ public class LiveServiceImpl implements LiveService {
     /**
      * sessionId 중복을 체크하는 메서드
      */
-    //url 중복 체크할 메서드
+    // url 중복 체크할 메서드
     public boolean getLiveCheckSessionIdBySessionId(String sessionId) {
         // 디비에 방송 url 정보 조회
         Optional<Live> oLive = liveRepository.findBySessionId(sessionId);
@@ -85,19 +82,14 @@ public class LiveServiceImpl implements LiveService {
     @Override
     public boolean postLiveByThumbnailUrl(Long liveId, String thumbnailUrl) {
 
-        List<Live> liveList = liveRepository.findAllById(liveId);
+        Optional<Live> oLive = liveRepository.findById(liveId);
+        Live live = oLive.orElseThrow(() -> new CustomException(LIVE_NOT_FOUND));
 
-        if (liveList == null || liveList.isEmpty())
-            throw new CustomException(LIVE_NOT_FOUND);
+        if (!live.isLive())
+            throw new CustomException(ALREADY_LIVE_END);
 
-        for (Live live : liveList) {
-            //현재 라이브를 하고 있을 때만 썸네일 바꾸기
-            if (live.isLive()) {
-                live.setThumbnailUrl(thumbnailUrl);
-                liveRepository.save(live);
-            }
-        }
-
+        live.setThumbnailUrl(thumbnailUrl);
+        liveRepository.save(live);
 
         return true;
     }
@@ -105,16 +97,14 @@ public class LiveServiceImpl implements LiveService {
     @Override
     public boolean postLiveByCategories(LiveCategoriesReq liveCategoriesReq) {
         Optional<Live> oLive = liveRepository.findById(liveCategoriesReq.getLiveId());
-        Live live = oLive.orElseThrow(()-> new CustomException(LIVE_NOT_FOUND));
+        Live live = oLive.orElseThrow(() -> new CustomException(LIVE_NOT_FOUND));
 
-        if(!live.isLive()){
+        if (!live.isLive()) {
             throw new CustomException(ALREADY_LIVE_END);
         }
 
-
-        //현재 라이브를 하고 있을 때만 카테고리 넣어주기
-        if (live.isLive()) {
-            live.getLiveCategoryList();
+        // 현재 라이브를 하고 있을 때만 카테고리 넣어주기
+        else {
             List<LiveCategory> liveCategoryList = new ArrayList<>();
 
             for (LiveCategoryReq liveCategoryReq : liveCategoriesReq.getLiveCategoryReqList()) {
@@ -128,7 +118,6 @@ public class LiveServiceImpl implements LiveService {
 
             }
             liveCategoryRepository.saveAll(liveCategoryList);
-
 
             liveRepository.save(live);
         }
@@ -147,16 +136,14 @@ public class LiveServiceImpl implements LiveService {
         for (Live live : liveList) {
             if (!live.isLive())
                 continue;
-            //라이브 카테고리 헬퍼 테이블 순회
+            // 라이브 카테고리 헬퍼 테이블 순회
             List<LiveCategory> liveCategories = live.getLiveCategoryList();
             List<Category> categoryList = new ArrayList<>();
 
+            for (LiveCategory liveCategory : liveCategories) {
+                // 카테고리 아이디와 연관된 카테고리 테이블 조회
 
-            for (Iterator<LiveCategory> it = liveCategories.iterator(); it.hasNext(); ) {
-                LiveCategory liveCategory = it.next();
-                //카테고리 아이디와 연관된 카테고리 테이블 조회
-
-                if(liveCategory.getCategory() == null){
+                if (liveCategory.getCategory() == null) {
                     throw new CustomException(CATEGORY_NOT_FOUND);
                 }
 
@@ -165,9 +152,9 @@ public class LiveServiceImpl implements LiveService {
                         .name(liveCategory.getCategory().getName())
                         .build());
             }
-            //라이브 카테고리 조회 끝
+            // 라이브 카테고리 조회 끝
 
-            //Live와 연관된 유저 조회
+            // Live 와 연관된 유저 조회
 
             List<UserLive> userLiveList = userLiveRepository.findAllByLive_id(live.getId());
 
@@ -196,7 +183,7 @@ public class LiveServiceImpl implements LiveService {
     public boolean postUserLiveByLiveId(LiveUserJoinReq liveUserJoinReq) {
         if (liveUserJoinReq.getUserId() == null || liveUserJoinReq.getLiveId() == null)
             throw new CustomException(INVALID_PARAMETER);
-        //유저 라이브 참가
+        // 유저 라이브 참가
         Optional<Live> oLive = liveRepository.findById(liveUserJoinReq.getLiveId());
         Live live = oLive
                 .orElseThrow(() -> new CustomException(LIVE_NOT_FOUND));
@@ -205,18 +192,17 @@ public class LiveServiceImpl implements LiveService {
         User user = oUser
                 .orElseThrow(() -> new CustomException(USER_NOT_FOUND));
 
-        List<UserLive> userLiveList = userLiveRepository.findAllByUser_idAndLive_id(liveUserJoinReq.getUserId(), liveUserJoinReq.getLiveId());
+        List<UserLive> userLiveList = userLiveRepository.findAllByUser_idAndLive_id(liveUserJoinReq.getUserId(),
+                liveUserJoinReq.getLiveId());
 
-        //유저 라이브 중복이 있다면
+        // 유저 라이브 중복이 있다면
         if (!userLiveList.isEmpty())
             throw new CustomException(ALREADY_SAVED_USER);
 
-        //이미 죽어있는 방송이면 들어가면 안됨
+        // 이미 죽어있는 방송이면 들어가면 안됨
         if (!live.isLive()) {
-            throw new CustomException(LIVE_NOT_FOUND);
+            throw new CustomException(ALREADY_LIVE_END);
         }
-        //라이브에 이미 들어가있으면 들어가면 안됨
-
 
         UserLive userLive = UserLive.builder()
                 .user(user)
@@ -233,8 +219,9 @@ public class LiveServiceImpl implements LiveService {
      */
     @Override
     public boolean deleteUserLiveByLiveId(LiveUserJoinReq liveUserJoinReq) {
-        //라이브 아이디 조회
-        List<UserLive> userLiveList = userLiveRepository.findAllByUser_idAndLive_id(liveUserJoinReq.getUserId(), liveUserJoinReq.getLiveId());
+        // 라이브 아이디 조회
+        List<UserLive> userLiveList = userLiveRepository.findAllByUser_idAndLive_id(liveUserJoinReq.getUserId(),
+                liveUserJoinReq.getLiveId());
 
         if (userLiveList == null || userLiveList.isEmpty())
             throw new CustomException(USER_LIVE_NOT_FOUND);
@@ -251,20 +238,21 @@ public class LiveServiceImpl implements LiveService {
 
     @Override
     public boolean patchLiveEndById(Long liveId) {
-        //라이브 상태 false로 변경
+        // 라이브 상태 false 로 변경
         Optional<Live> oLive = liveRepository.findById(liveId);
         Live live = oLive.orElseThrow(() -> new CustomException(LIVE_NOT_FOUND));
 
+        if (!live.isLive())
+            throw new CustomException(ALREADY_LIVE_END);
         live.setLive(false);
 
-        //라이브에 찜한 유저의 찜 테이블에서 전부 삭제
+        // 라이브에 찜한 유저의 찜 테이블에서 전부 삭제
         List<Favorite> favoriteList = favoriteRepository.findByLive_id(liveId);
         favoriteRepository.deleteAll(favoriteList);
-        //라이브에 참가한 유저, 유저라이브 테이블에서 전부 삭제
-        //라이브 아디와 유저 아디가 판매자 본인과 만났을 경우 라이브 유저 전부 삭제해서 주석처리
+        // 라이브에 참가한 유저, 유저라이브 테이블에서 전부 삭제
+        // 라이브 아디와 유저 아디가 판매자 본인과 만났을 경우 라이브 유저 전부 삭제해서 주석처리
         List<UserLive> userLiveList = userLiveRepository.findAllByLive_id(liveId);
         userLiveRepository.deleteAll(userLiveList);
-
 
         liveRepository.save(live);
         return true;
@@ -281,7 +269,7 @@ public class LiveServiceImpl implements LiveService {
         for (LiveContent liveContent : liveContentList) {
             for (Category tempCategory : liveContent.getCategories()) {
                 if (tempCategory.getName().equals(category)) {
-                    //해당 카테고리에 값과 맞다면
+                    // 해당 카테고리에 값과 맞다면
                     tempLiveContentList.add(liveContent);
                     break;
                 }
@@ -289,18 +277,18 @@ public class LiveServiceImpl implements LiveService {
 
         }
 
-
         return tempLiveContentList;
     }
 
     /**
      * @param liveContentList 제목, 카테고리 등으로 필터링된 값
      * @param location        현재 구매자의 위치
-     * @param isNational      true면 전국, false면 지역 5km 이내
+     * @param isNational      true 면 전국, false 면 지역 5km 이내
      * @return 필터링되고 정렬된 값
      */
     @Override
-    public List<DistanceModule> searchLocationLiveList(List<LiveContent> liveContentList, Location location, boolean isNational) {
+    public List<DistanceModule> searchLocationLiveList(List<LiveContent> liveContentList, Location location,
+            boolean isNational) {
         List<DistanceModule> distanceModuleList = new ArrayList<>();
         for (LiveContent liveContent : liveContentList) {
             // 라이브 아이디로 lat,lon 조회
@@ -312,14 +300,13 @@ public class LiveServiceImpl implements LiveService {
             if (location.getLatitude() == null || location.getLongitude() == null)
                 throw new CustomException(INVALID_PARAMETER);
             // 킬로미터(Kilo Meter) 단위
-            double distanceKiloMeter =
-                    LocationDistance.distance(location.getLatitude(), location.getLongitude(),
-                            live.getLatitude(), live.getLongitude(), "kilometer");
+            double distanceKiloMeter = LocationDistance.distance(location.getLatitude(), location.getLongitude(),
+                    live.getLatitude(), live.getLongitude(), "kilometer");
 
-            //전국이면
+            // 전국이면
             if (isNational) {
                 distanceModuleList.add(new DistanceModule(distanceKiloMeter, liveContent));
-            } else {//전국이 아니면 5km 이내
+            } else {// 전국이 아니면 5km 이내
                 if (distanceKiloMeter <= 5) {
                     distanceModuleList.add(new DistanceModule(distanceKiloMeter, liveContent));
                 }
@@ -337,22 +324,12 @@ public class LiveServiceImpl implements LiveService {
      */
     @Override
     public List<LiveContent> searchSortUserJoinLiveList(List<LiveContent> liveContentList, String userJoinSort) {
-        //오름차순 정렬
+        // 오름차순 정렬
         if (userJoinSort.equals("ASC")) {
-            Collections.sort(liveContentList, new Comparator<LiveContent>() {
-                @Override
-                public int compare(LiveContent o1, LiveContent o2) {
-                    return o1.getJoinUsersNum() - o2.getJoinUsersNum();
-                }
-            });
-            //내림차순 정렬
+            liveContentList.sort((o1, o2) -> o1.getJoinUsersNum() - o2.getJoinUsersNum());
+            // 내림차순 정렬
         } else if (userJoinSort.equals("DESC")) {
-            Collections.sort(liveContentList, new Comparator<LiveContent>() {
-                @Override
-                public int compare(LiveContent o1, LiveContent o2) {
-                    return o2.getJoinUsersNum() - o1.getJoinUsersNum();
-                }
-            });
+            liveContentList.sort((o1, o2) -> o2.getJoinUsersNum() - o1.getJoinUsersNum());
 
         }
         return liveContentList;
@@ -387,24 +364,22 @@ public class LiveServiceImpl implements LiveService {
      * @param liveId 조회할 라이브 아이디
      * @return 라이브 상세보기 값
      */
-    //방 상세보기 가져올 메서드
+    // 방 상세보기 가져올 메서드
     @Override
     public LiveDetailGetRes getLiveDetailBySessionId(Long liveId) {
         // 디비에 방송 url 정보 조회
         Optional<Live> oLive = liveRepository.findById(liveId);
         Live live = oLive.orElseThrow(() -> new CustomException(LIVE_NOT_FOUND));
 
-        if(live == null)
+        if (live == null)
             return null;
 
-        //라이브 카테고리 헬퍼 테이블 순회
+        // 라이브 카테고리 헬퍼 테이블 순회
         List<LiveCategory> liveCategories = live.getLiveCategoryList();
         List<Category> categoryList = new ArrayList<>();
 
-
-        for (Iterator<LiveCategory> it = liveCategories.iterator(); it.hasNext(); ) {
-            LiveCategory liveCategory = it.next();
-            //카테고리 아이디와 연관된 카테고리 테이블 조회
+        for (LiveCategory liveCategory : liveCategories) {
+            // 카테고리 아이디와 연관된 카테고리 테이블 조회
             if (liveCategory.getCategory() == null)
                 continue;
             categoryList.add(Category.builder()
@@ -412,18 +387,16 @@ public class LiveServiceImpl implements LiveService {
                     .name(liveCategory.getCategory().getName())
                     .build());
         }
-        //라이브 카테고리 조회 끝
+        // 라이브 카테고리 조회 끝
 
-        //Live와 연관된 유저 조회
+        // Live 와 연관된 유저 조회
         User user = live.getUser();
-
 
         List<UserLive> userLiveList = userLiveRepository.findAllByLive_id(live.getId());
         List<UserEntryRes> userEntryList = new ArrayList<>();
 
-        for (Iterator<UserLive> it = userLiveList.iterator(); it.hasNext(); ) {
-            UserLive userLive = it.next();
-            //유저 아이디와 연관된 유저 테이블 조회
+        for (UserLive userLive : userLiveList) {
+            // 유저 아이디와 연관된 유저 테이블 조회
 
             UserEntryRes userEntryRes = UserEntryRes.builder()
                     .id(userLive.getUser().getId())
@@ -432,11 +405,13 @@ public class LiveServiceImpl implements LiveService {
 
             userEntryList.add(userEntryRes);
         }
-        //라이브 아이디와 연관된 상품 테이블 조회
+        // 라이브 아이디와 연관된 상품 테이블 조회
         Optional<List<Product>> oProduct = productRepository.findByLive_Id(live.getId());
         List<Product> productList = oProduct.orElse(null);
 
         List<LiveProductInfo> liveProductInfoList = new ArrayList<>();
+        //productList 가 없을수도 있으니
+        assert productList != null;
         for (Product product : productList) {
 
             LiveProductInfo liveProductInfo = LiveProductInfo.builder()
@@ -460,9 +435,9 @@ public class LiveServiceImpl implements LiveService {
             liveProductInfoList.add(liveProductInfo);
         }
 
+        // 불러온 값 넣어주기
 
-        //불러온 값 넣어주기
-        LiveDetailGetRes liveDetailGetRes = LiveDetailGetRes.builder()
+        return LiveDetailGetRes.builder()
                 .id(live.getId())
                 .categories(categoryList)
                 .seller_id(user.getId())
@@ -475,9 +450,6 @@ public class LiveServiceImpl implements LiveService {
                 .userEntryResList(userEntryList)
                 .liveProductInfoList(liveProductInfoList)
                 .build();
-
-
-        return liveDetailGetRes;
 
     }
 }
